+++ src/app/eth-common/module/toolbar/alethioCompanyModule.tsx
@@ -2,7 +2,7 @@ import React from "react";
 import { IModuleDef } from "plugin-api/IModuleDef";
 import { ToolbarItem } from "@alethio/ui/lib/layout/toolbar/ToolbarItem";
 import { ITranslation } from "plugin-api/ITranslation";
-import { ExternalLink } from "plugin-api/component/ExternalLink";
+import { Link } from "plugin-api/component/Link";
 import { ToolbarIconButton } from "@alethio/ui/lib/layout/toolbar/ToolbarIconButton";
 import { BriefcaseIcon } from "@alethio/ui/lib/icon/BriefcaseIcon";
 
@@ -16,9 +16,9 @@ export const alethioCompanyModule: (alethioCompanyUrl: string) => IModuleDef<IAl
     dataAdapters: [],
     getContentComponent: async () => (props) => (
         <ToolbarItem title={props.translation.get("toolbar.company.label")}>
-            <ExternalLink href={alethioCompanyUrl} rel="noopener noreferrer">
+            <Link to={alethioCompanyUrl}>
                 <ToolbarIconButton Icon={BriefcaseIcon} />
-            </ExternalLink>
+            </Link>
         </ToolbarItem>
     ),
     getContentProps : data => data,